moving around some report card code and working on app capabilities work
